pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"



[project]
name = "rag-project"
version = "0.1.0"
description = "RAG system with functional programming approach"
readme = "README.md"
requires-python = ">=3.11"  
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi>=0.119.0",
    "google-generativeai>=0.8.5",
    "langchain-huggingface>=0.3.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "pypdf2>=3.0.1",
    "pytest>=8.4.2",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.2.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.1",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn[standard]>=0.37.0",
]



# Hatchling configuration for src layout
[tool.hatch.build.targets.wheel]
packages = ["src/rag"]

#[tool.hatch.build.targets.wheel.shared-data]
#"data" = "share/rag/data"